* created src/ folder containing neurostatslib * moved package scripts into src * removed setup.py and replaced installation with pyproject.toml
