+++ pyproject.toml
@@ -1,0 +1,51 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "neurostatslib"
+
+version = "0.1.0"
+
+authors = [
+    {name = "Edoardo Balzani", email = "ebalzani@flatironinstitute.org"},
+    {name = "William Broderick", email = "wbroderick@flatironinstitute.org"},
+    {name = "Guillaume Vejo", email = "gviejo@flatironinstitute.org"},
+    {name = "Alex Williams", email = "alex.h.williams@nyu.edu"}
+]
+description = "Toolbox for basic Generalized Linear Models (GLMs) for neural data analysis"
+
+readme = "README.md"
+
+requires-python = ">=3.7"
+
+keywords = ["neuroscience", "Poisson-GLM"]
+
+license = {file = "LICENSE"}
+
+classifiers = [
+    "Intended Audience :: Science/Research",
+    "Natural Language :: English",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10"
+]
+
+dependencies = [
+    'jax>=0.4',
+    'jaxopt>=0.6',
+    'matplotlib>=3.7',
+    # for numpy.typing.ArrayLike
+    'numpy>1.20',
+    'scikit-learn>=1.2',
+    'scipy>=1.10',
+    'typing_extensions>=4.6'
+]
+
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["neurostatslib"]
+
