+++ src/neurostatslib/glm.py
